refactor: extract about overlay into ui/overlays/about.rs (#357)
Sets up the ui/overlays/ subdirectory with one file per overlay,
following the issue's proposed structure. First migrant is the
about overlay -- smallest, fully self-contained, just renders
version + author + repo link in a centered popup.

To support future overlay migrations, marked centered_popup and
ABOUT_POPUP_WIDTH as pub(super) so submodules under ui/overlays/
can reuse them. Subsequent overlay extractions will follow the
same pattern; centered_popup may eventually move into
ui/overlays/mod.rs once all overlays have migrated.

ui/mod.rs: 4849 -> 4800 lines.
